requirements.txt
colorama==0.4.6
greenlet==3.0.3
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
tqdm==4.66.2
typing_extensions==4.9.0